requirements.txt
python-dotenv
flask
sqlalchemy